feat: rename components
