+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next"; // Object to set metadata
 import { Roboto } from "next/font/google"; // You can change the font to anything you want.
 import "./globals.css";
-import Navbar from "@/Components/navbar";
-import Footer from "@/Components/footer";
+import Navbar from "@/components/navbar";
+import Footer from "@/components/footer";
 
 // If you are experiencing an error "localFont is undefined", you might need to add the following blocks of code
 const roboto = Roboto({ subsets: ["latin"] });
