+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Homepage from "@/Components/homepage";
+import Homepage from "@/components/homepage";
 export default function Home() {
   return (
 		<div>
