+++ src/components/projectPreview.tsx
